+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- [#298](https://github.com/deviceinsight/kafkactl/issues/298) Support filtering ACL deletion by resource name
+
 ## 5.20.0 - 2026-07-30
 
 ### Changed